refactor(memory): split store.py (24 KB → 12 KB) along table boundaries
CLAUDE.md rule #5 (모듈 ≤ 20 KB) 위반 해소. core/memory/store.py 가
576줄 / 24 KB 까지 커져 있어 자연 경계 4개로 분할.

분할 결과:
- core/memory/db.py          (3.5 KB) DB_PATH / _connect / init_db / persona flag
- core/memory/conversation.py(6.5 KB) save_turn / get_recent_turns /
                                       get_history_context / get_all_sessions /
                                       set_session_name / delete_session
- core/memory/summaries.py   (3 KB)   save_session_summary /
                                       get_session_summaries / get_long_term_context
- core/memory/store.py       (12 KB)  MemoryStore facade — insights /
                                       persona / get_stats 유지 + 위 모듈로 위임

공개 API 보존:
- `from core.memory.store import MemoryStore, DB_PATH, _connect, init_db`
  모두 그대로 동작 (re-export). server_llmwiki.py / character_profile.py /
  knowledge_tracker.py / tests 등 17 호출 측 변경 없음.
- MemoryStore 메서드 시그니처/반환값 동일. 13개 호출 site (.save/.save_turn/
  .get_persona 등) 무변경.

테스트 갱신 (tests/test_conversation_continuity.py):
- StoreTruncationTests: 소스 introspection 대상을 core.memory.conversation
  으로 이전 (실제 [:2000] / [:800] 캡이 그쪽으로 이동).
- StoreRoundTripTests: DB_PATH monkey-patch 를 core.memory.db 에 적용 —
  _connect() 가 그쪽에서 dynamically 읽으므로 그게 진짜 source of truth.

검증:
- python -m pytest tests/ → 1714 passed, 0 failed, 350 subtests OK
- 직접 facade smoke: save/get_stats/persona/save_turn/get_recent_turns/
  get_history_context/get_all_sessions/set_session_name/
  save_session_summary/get_session_summaries 모두 동작 확인

ROADMAP v0.3 carryover follow-up "core/memory/store.py split" 항목 해소.
docs/handovers/v0.3.0-platform-track.md §3 사이클 6 첫 항목.
